Refactoring- moved logic in controller to services, renamed desired_occupation column to information, rearranged files and endpoints
