+++ src/main/java/com/tucklets/app/controllers/BasicController.java
@@ -33,7 +33,6 @@ public class BasicController {
         admin.setUsername("admin");
         admin.setEnabled(true);
         adminService.saveAdmin(admin);
-        return "redirect:/admin/dashboard";
+        return "redirect:/admin/children-dashboard/";
     }
-
 }
