+++ src/main/resources/i18n/messages_en.properties
@@ -23,5 +23,5 @@ sponsor-a-child.subtitle=Your donation will keep one child in school and provide
 sponsor-a-child.description=You can sponsor one or more children or teacher from the list below.
 child.age=Age:\u0020
 child.grade=Grade:\u0020
-child.info=Aspirations:\u0020
+child.info=Information:\u0020
 sponsor-a-child.submit=Sponsor Now
