+++ src/main/resources/static/css/admin/dashboard.css
@@ -35,7 +35,7 @@
     border: 0;
 }
 /* The Modal box */
-.modify-child-modal {
+.modal-div {
     background-color: #fefefe;
     margin: auto;
     padding: 20px;
