+++ src/main/resources/static/js/admin/upload-children-data.js
@@ -1,0 +1,3 @@
+function handleUploadFile() {
+    $('.upload-button').prop('disabled', !$('.upload-file-button').val());
+}
