--- src/main/resources/static/js/upload-children-data.js
@@ -1,3 +1,0 @@
-function handleUploadFile() {
-    $('.upload-button').prop('disabled', !$('.upload-file-button').val());
-}
