+++ src/main/resources/templates/admin/admin-nav-bar.html
@@ -1,8 +1,8 @@
 <div class="admin-navigation"  th:fragment="admin-nav-bar">
     <ul class="nav-bar">
-        <li class="nav-element"><a th:href="@{/admin/dashboard}" class="nav-link">Students</a></li>
-        <li class="nav-element"><a th:href="@{/sponsor-a-child/}" class="nav-link">Sponsors</a></li>
-        <li class="nav-element"><a th:href="@{/admin/dashboard/upload}" class="nav-link">Upload Children</a></li>
+        <li class="nav-element"><a th:href="@{/admin/children-dashboard/}" class="nav-link">Children</a></li>
+        <li class="nav-element"><a th:href="@{/admin/sponsor-dashboard/}" class="nav-link">Sponsors</a></li>
+        <li class="nav-element"><a th:href="@{/admin/children-dashboard/upload}" class="nav-link">Upload Children</a></li>
         <li class="nav-element-right"><a th:href="@{/logout}" class="nav-link">Logout</a></li>
     </ul>
 </div>
