+++ src/main/resources/templates/admin/child-modal.html
@@ -1,5 +1,5 @@
 <div class="modal-container" id="modal-container" th:fragment="modify-child-modal">
-  <div class="modify-child-modal form-group">
+  <div class="modal-div form-group">
     <div class="modal-header">
       <div>
         <h5 class="modal-title" id="edit_title">Edit a child</h5>
@@ -9,7 +9,7 @@
         <span aria-hidden="true">×</span>
       </button>
     </div>
-    <form class="modify-child-form" th:action="@{/admin/dashboard/modify-child}" th:object="${childDetails}" method="post" enctype="multipart/form-data">
+    <form class="modify-child-form" th:action="@{/admin/children-dashboard/submit-child}" th:object="${childDetails}" method="post" enctype="multipart/form-data">
       <fieldset>
         <input type="hidden" id="child-id" th:field="*{child.childId}" />
         <input type="file" th:id="|upload-image-*{child.childId}|" class="icon-image-input" th:onchange="|uploadChildImage('${childDetails.child.childId}')|" th:field="*{childImageFile}">
@@ -20,8 +20,8 @@
         <input type="text" id="child-last-name" class="form-control" th:field="*{child.lastName}"/>
         <label for="child-birth-year">Birth year:</label>
         <input type="text" id="child-birth-year" class="form-control" th:field="*{child.birthYear}"/>
-        <label for="child-desired-occupation">Desired Occupation:</label>
-        <textarea  id="child-desired-occupation" class="form-control" th:field="*{child.desiredOccupation}"></textarea>
+        <label for="child-information">Information:</label>
+        <textarea  id="child-information" class="form-control" th:field="*{child.information}"></textarea>
         <label for="child-grade">Grade:</label>
         <input type="text" id="child-grade" class="form-control" th:field="*{child.grade}" />
         <label for="sponsored">Sponsored:</label>
