+++ src/main/resources/templates/admin/children-dashboard.html
@@ -1,0 +1,70 @@
+<!DOCTYPE HTML>
+<html lang="en" xmlns:th="http://www.w3.org/1999/xhtml">
+<head>
+    <div th:replace="common-head :: common-head"></div>
+    <link rel="stylesheet" th:href="@{/css/admin/dashboard.css}">
+    <link rel="stylesheet" th:href="@{/css/admin/admin-common.css}">
+    <script type="text/javascript" th:src="@{/js/admin/children-dashboard.js}"></script>
+
+</head>
+
+<body>
+<div>
+    <div th:replace="header :: header" />
+    <div th:replace="admin/admin-nav-bar :: admin-nav-bar" />
+
+    <div class="jumbotron jumbotron-fluid">
+        <div class="container">
+            <h1 class="display-4">Admin Dashboard</h1>
+            <p class="lead">View information about children, sponsors, sponsorships, etc.</p>
+        </div>
+    </div>
+    <div class="children-table">
+        <h4>Children Information</h4>
+        <table class="table table-striped table-bordered children-data-table">
+            <thead>
+            <tr>
+                <th></th>
+                <th scope="col">Image</th>
+                <th scope="col">Child Id</th>
+                <th scope="col">First Name</th>
+                <th scope="col">Last Name</th>
+                <th scope="col">Birth Year</th>
+                <th scope="col">Grade</th>
+                <th scope="col">Is Sponsored?</th>
+                <th scope="col">Information</th>
+                <th scope="col">Created</th>
+                <th scope="col">Last Updated</th>
+                <th scope="col">Action</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr th:each="childContainer : ${childrenDetails}" th:id="child-row + ${childContainer.child.childId}">
+                <td><input type="image" class="delete-child-button" src="/images/red-x.jpg" th:onclick="|handleDeleteChild('${childContainer.child.childId}')|" /> </td>
+                <td><img class="icon-image" th:src="${childContainer.childImageLocation}" /></td>
+                <td th:text="*{childContainer.child.childId}"/>
+                <td th:text="*{childContainer.child.firstName}"/>
+                <td th:text="*{childContainer.child.lastName}"/>
+                <td th:text="*{childContainer.child.birthYear}"/>
+                <td th:text="*{childContainer.child.grade}"/>
+                <td th:text="*{childContainer.child.sponsored}"/>
+                <td th:text="*{childContainer.child.information}"/>
+                <td th:text="*{childContainer.child.creationDate}"/>
+                <td th:text="*{childContainer.child.lastUpdateDate}"/>
+                <td>
+                    <input type="button" class="btn btn-primary" value="Edit" th:onclick="|handleEditChildClick('${childContainer.child.childId}')|" />
+                </td>
+            </tr>
+            </tbody>
+        </table>
+    </div>
+    <div class="admin-tools-div">
+        <a th:href="@{/admin/children-dashboard/upload}" class="btn btn-primary upload-children-data-button">Upload Children Data</a>
+        <input type="button" class="btn btn-primary" value="Add Child" onclick="handleAddChildClick()" />
+        <input type="button" id="export-children-button" class="btn btn-primary" onclick="exportChildren()" value="Export" />
+    </div>
+    <div id="modify-child-modal-holder"></div>
+</div>
+</body>
+</html>
+
