+++ src/main/resources/templates/header.html
@@ -6,7 +6,7 @@
       <a class="navbar-brand" href="#">
         <img class="d-inline-block align-top tucklets-logo" alt="" th:src="@{/images/tucklets-icon.PNG}"/>
       </a>
-      <select class="supported-languages-dropdown" th:if="${supportedLocales != null}">
+      <select class="supported-languages-dropdown" th:if="${localeContainer.supportedLocales != null}">
         <option th:each="locale : ${localeContainer.supportedLocales}" th:value="${locale.toString()}" th:text="${locale.displayName}" th:selected="${locale} == ${localeContainer.selectedLocale}"/>
       </select>
     </nav>
