--- src/main/resources/templates/login.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.w3.org/1999/xhtml">
-<head>
-    <div th:replace="common-head :: common-head"></div>
-    <link rel="stylesheet" th:href="@{/css/login.css}">
-</head>
-
-<div class="tucklets-login tucklets-login-background">
-    <div th:if="${param.error}">
-        Invalid username and password.
-    </div>
-    <div th:if="${param.logout}">
-        You have been logged out.
-    </div>
-    <div class="login-modal">
-        <div class="sign-in-title">
-            <span>Sign In </span>
-        </div>
-        <form th:action="@{/login}" class="login-form" method="post">
-            <label for="login-username">Username</label>
-            <input id="login-username" class="form-control" type="text" name="username" placeholder="Username">
-            <label for="login-password">Password</label>
-            <input id="login-password" class="form-control" type="password" name="password" placeholder="******">
-            <div class="login-submission-div">
-                <input type="submit" class="btn btn-primary" value="Login">
-            </div>
-        </form>
-    </div>
-</div>
-</html>
-
-
