+++ src/main/resources/templates/sponsor-a-child.html
@@ -21,7 +21,7 @@
     <form class="select-child-form container" th:action="@{/sponsor-info/}" method="get">
       <div class="container children-div">
         <label
-                th:each="childContainer : ${sponsorChildrenContainer.children}"
+                th:each="childContainer : ${sponsorAChildContainer.children}"
                 th:id="child-id + ${childContainer.child.childId}"
                 th:for="checkbox- + ${childContainer.child.childId}"
                 class="card child-div"
@@ -33,7 +33,7 @@
             <h5 class="card-title" th:text="|${childContainer.child.firstName} ${childContainer.child.lastName}|" />
             <p class="card-text"><b th:text="#{child.age}" /><span th:text="${childContainer.age}"/></p>
             <p class="card-text"><b th:text="#{child.grade}" /><span th:text="${childContainer.child.grade}"/></p>
-            <p class="card-text"><b th:text="#{child.info}" /><span th:text="${childContainer.child.desiredOccupation}"/></p>
+            <p class="card-text"><b th:text="#{child.info}" /><span th:text="${childContainer.child.information}"/></p>
             <input type="checkbox" name="childId" th:id="checkbox- + ${childContainer.child.childId}" th:value="${childContainer.child.childId}">
           </div>
         </label>
